removes buttons when cards end #52 :bug:
